:sparkles: feature:添加用户、人才和负责人登录注册代码业务

+++ src/main/java/com/voyager/mapper/UserMapper.java
@@ -14,7 +14,7 @@ public interface UserMapper {
      * @return 对应的User对象
      */
     @Select("SELECT * FROM User WHERE user_id = #{userId}")
-    User findByUserId(int userId);
+    User findByUserId(Long userId);
 
     /**
      * 根据用户名查询用户
